+++ services/index.js
@@ -2,10 +2,12 @@ const advertise = require('./advertise');
 const openBalancedChannel = require('./open_balanced_channel');
 const purchasePing = require('./purchase_ping');
 const serviceKeySendRequests = require('./service_key_send_requests');
+const simulateKeySendRequest = require('./simulate_key_send_request');
 
 module.exports = {
   advertise,
   openBalancedChannel,
   purchasePing,
   serviceKeySendRequests,
+  simulateKeySendRequest,
 };